requirements.txt
torch
torchvision
torchaudio
segmentation-models-pytorch
albumentations
opencv-python
numpy
matplotlib
tqdm
pyyaml
